requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cmake==3.22.3
colorama==0.4.4
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dnspython==2.2.1
email-validator==1.1.3
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.6.0
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.0
numpy==1.22.3
Pillow==9.0.1
psycopg2==2.9.3
py3-validate-email==1.0.5
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
twilio==7.8.0
tzdata==2021.5
urllib3==1.26.9
waitress==2.1.1
